Restructure folders to use the new format
