+++ src/TanachStudy/Parts/PartList.jsx
