+++ src/TanachStudy/Parts/parts.json
