+++ src/TanachStudy/Perakim/TeacherCard.jsx
