+++ src/TanachStudy/Perakim/TeamimCard.jsx
