+++ src/TanachStudy/Perakim/TeamimItem.css
