+++ src/TanachStudy/Perakim/TeamimItem.jsx
