+++ src/TanachStudy/PerekList/NachPerekItem/NachPerekItem.jsx
