+++ src/TanachStudy/PerekList/TorahPerekItem/TorahPerekItem.jsx
