+++ src/TanachStudy/Teachers/Teacher.css
